+++ ReactNativeCore/.expo/packager-info.json
@@ -2,8 +2,8 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 12672,
+  "packagerPid": 10936,
   "expoServerNgrokUrl": "https://3p-299.dderrickmatheww.reactnativecore.exp.direct",
   "packagerNgrokUrl": "https://packager.3p-299.dderrickmatheww.reactnativecore.exp.direct",
-  "ngrokPid": 19940
+  "ngrokPid": 9364
 }
