+++ ReactNativeCore/package.json
@@ -11,11 +11,14 @@
     "@react-native-community/masked-view": "^0.1.7",
     "@react-navigation/bottom-tabs": "^5.2.5",
     "@react-navigation/native": "^5.1.4",
+    "@react-navigation/stack": "^5.2.10",
     "axios": "^0.19.2",
     "enzyme": "^3.11.0",
     "expo": "~36.0.0",
     "expo-constants": "^9.0.0",
     "expo-facebook": "^8.1.0",
+    "expo-google-app-auth": "^8.1.0",
+    "expo-google-sign-in": "^8.1.0",
     "firebase": "^7.13.2",
     "firebase-app": "0.0.1",
     "firebase-auth": "^0.1.2",
